requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pymongo==3.12.1
sqlparse==0.3.1
python-dotenv==1.0.0
django-cors-headers==4.3.0
pyjwt==2.8.0
bcrypt==4.0.1
pillow==10.0.0
gunicorn==21.2.0 